修正3

+++ app/controllers/items_controller.rb
@@ -37,9 +37,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item = Item.find(params[:id])
+      item.user != current_user
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
